refactor: types structure
