+++ src/balances/accounts/types.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::types::FormattedAmount;
+use crate::types::internal::FormattedAmount;
 
 #[derive(Default, Debug, Serialize)]
 pub struct ParsedAta {
